# api-docs.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Backend Santander Stock Screener",
        "description": "Backend for the clone of the Santander stock screener website",
        "termsOfService": "http://swagger.io/terms",
        "license": {
            "name": "Apache 2.0",
            "url": "http://springdoc.org"
        },
        "version": "1.0.0"
    },
    "servers": [{
        "url": "http://localhost:8080",
        "description": "Generated server url"
    }],
    "paths": {
        "/stock": {
            "get": {
                "tags": ["stock-controller"],
                "operationId": "findAll",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/StockDTO"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": ["stock-controller"],
                "operationId": "update",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StockDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StockDTO"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": ["stock-controller"],
                "operationId": "save",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StockDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StockDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/stock/{id}": {
            "get": {
                "tags": ["stock-controller"],
                "operationId": "findById",
                "parameters": [{
                    "name": "id",
                    "in": "path",
                    "required": true,
                    "schema": {
                        "type": "integer",
                        "format": "int64"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StockDTO"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": ["stock-controller"],
                "operationId": "delete",
                "parameters": [{
                    "name": "id",
                    "in": "path",
                    "required": true,
                    "schema": {
                        "type": "integer",
                        "format": "int64"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StockDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/stock/today": {
            "get": {
                "tags": ["stock-controller"],
                "operationId": "findByToday",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/StockDTO"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "StockDTO": {
                "required": ["date", "name", "price", "variation"],
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string"
                    },
                    "price": {
                        "minimum": 0.00,
                        "exclusiveMinimum": false,
                        "type": "number",
                        "format": "double"
                    },
                    "date": {
                        "type": "string",
                        "format": "date"
                    },
                    "variation": {
                        "type": "number",
                        "format": "double"
                    }
                }
            }
        }
    }
}