# src/main/java/com/pedrogobira/santanderhomebrokerbackend/model/dto/StockDTO.java
package com.pedrogobira.santanderhomebrokerbackend.model.dto;

import com.fasterxml.jackson.annotation.JsonFormat;

import javax.validation.constraints.DecimalMin;
import javax.validation.constraints.Digits;
import javax.validation.constraints.NotNull;
import java.time.LocalDate;

public class StockDTO {

	private Long id;

	@NotNull
	private String name;

	@NotNull
	@DecimalMin(value = "0.00")
	@Digits(integer = 6, fraction = 2)
	private Double price;

	@NotNull
	@JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "dd/MM/yyyy")
	private LocalDate date;

	@NotNull
	@Digits(integer = 3, fraction = 2)
	private Double variation;

	public Long getId() {
		return this.id;
	}

	public void setId(Long id) {
		this.id = id;
	}

	public String getName() {
		return this.name;
	}

	public void setName(String name) {
		this.name = name;
	}

	public Double getPrice() {
		return this.price;
	}

	public void setPrice(Double price) {
		this.price = price;
	}

	public LocalDate getDate() {
		return this.date;
	}

	public void setDate(LocalDate date) {
		this.date = date;
	}

	public Double getVariation() {
		return this.variation;
	}

	public void setVariation(Double variation) {
		this.variation = variation;
	}

}